pyproject.toml
[project]
name = "ultimate-conversor-bot"
version = "0.1.0"
description = "Ultimate file conversor."
authors = [
    {name = "Alekssandher",email = "94875660+Alekssandher@users.noreply.github.com"}
]
license = {text = "GPL-3"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

